gpiolib: acpi: Move quirks to a separate file
[ Upstream commit 92dc572852ddcae687590cb159189004d58e382e ]

The gpiolib-acpi.c is huge enough even without DMI quirks.
Move them to a separate file for a better maintenance.

No functional change intended.

Acked-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Stable-dep-of: 2d967310c49e ("gpiolib: acpi: Add quirk for Dell Precision 7780")
(cherry picked from commit b2023685b2de4437f39e6af8d0cd524747384a53)
